requirements.txt
# FastAPI
fastapi==0.115.0
uvicorn[standard]==0.32.0

# PostgreSQL 데이터베이스
sqlalchemy==2.0.44
alembic==1.14.0
psycopg2-binary>=2.9.9       
asyncpg>=0.30.0           

# SQLite (개발/테스트용 - 선택)
aiosqlite==0.20.0

# 인증/보안
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9

# 캐시
redis==5.2.0

# 설정
pydantic-settings==2.6.1
python-dotenv==1.0.1

# HTTP 클라이언트
httpx==0.28.0

# AWS S3 (Lightsail Storage)
boto3==1.35.0

# 이미지 처리
pillow>=10.4.0